Create MBBSEmuIntegrationTest for Echo Tests.
Also some minor fixes to libc functions.
